Refactored music functionality into a music cog/extension
